Feat: Add assignment on KNN in Scikit-Learn done to container
In the assignment, two python objects were saved:
i. diamond_model.pkl: The model created in the assignment
ii. diamond_scaler.pkl: The Scaler (StandardScaler) for the model, created from fitting into the training set parameters

A note was also added on splitting datasets before scaling, to prevent data leakage (a situation where parameters from the test set is leaked into the training set and the models learns it)

We fit and transform on the training set, and then use the parameters learned from the training set to transform the test set